5_Multimodal_Sentiment_Analysis_Model/requirements.txt
torch==1.9.0
Pillow==9.1.1
scikit-learn==1.1.1
torchvision==0.10.0
transformers==4.19.2
tqdm==4.64.0
scipy==1.8.1
numpy==1.22.4
